* kozo/master:
  Dashboard API
  Require old password
  Permission checking
  Grant and revoke permissions

+++ users.go
@@ -27,6 +27,11 @@ func optionsUser(t *Task) {
 }
 
 func listUsers(t *Task) {
+	if !hasPermission(t.Tx, t.Uid, "GET", "users", "") {
+		t.Rw.WriteHeader(http.StatusForbidden)
+		return
+	}
+
 	whereClause1, whereClause2, params := "", "", []interface{}{}
 
 	gid := t.Rq.URL.Query().Get("group")
@@ -90,7 +95,7 @@ func listUsers(t *Task) {
 	rows, err = t.Tx.Query(`
 		SELECT "user_id", "method", "object_type", "object_id"
 		FROM "permissions"
-		JOIN "users_to_groups" USING ("group_id")`+
+		JOIN "users_to_groups" USING ("group_id")` +
 		whereClause2)
 	if err != nil {
 		panic(err)
@@ -127,6 +132,11 @@ func listUsers(t *Task) {
 }
 
 func createUser(t *Task) {
+	if !hasPermission(t.Tx, t.Uid, "POST", "users", "") {
+		t.Rw.WriteHeader(http.StatusForbidden)
+		return
+	}
+
 	data, ok := t.RecvJson().(map[string]interface{})
 	if !ok {
 		t.Rw.WriteHeader(http.StatusBadRequest)
@@ -187,6 +197,11 @@ func createUser(t *Task) {
 }
 
 func getUser(t *Task) {
+	if !hasPermission(t.Tx, t.Uid, "GET", "user", t.UUID) {
+		t.Rw.WriteHeader(http.StatusForbidden)
+		return
+	}
+
 	rows, err := t.Tx.Query(`
 		SELECT "id", "name", "email", "created"
 		FROM "users"
@@ -261,6 +276,11 @@ func getUser(t *Task) {
 }
 
 func changeUser(t *Task) {
+	if !hasPermission(t.Tx, t.Uid, "PATCH", "user", t.UUID) {
+		t.Rw.WriteHeader(http.StatusForbidden)
+		return
+	}
+
 	if !userExists(t.Tx, t.UUID) {
 		t.Rw.WriteHeader(http.StatusNotFound)
 		return
@@ -299,6 +319,19 @@ func changeUser(t *Task) {
 	}
 
 	if passwdStr, ok := data["password"].(string); ok {
+		row := t.Tx.QueryRow(`SELECT "password" FROM "users"
+			WHERE "id" = $1`, t.UUID)
+		var oldHash []byte
+		if err := row.Scan(&oldHash); err != nil {
+			panic(err)
+		}
+		oldPasswdStr, ok := data["oldPassword"].(string)
+		oldPasswd := []byte(oldPasswdStr)
+		if !ok || bcrypt.CompareHashAndPassword(oldHash, oldPasswd) != nil {
+			t.SendError("'oldPassword' is invalid")
+			return
+		}
+
 		if passwdStr == "" {
 			t.SendError("'password' is invalid")
 			return
@@ -326,6 +359,11 @@ func changeUser(t *Task) {
 }
 
 func deleteUser(t *Task) {
+	if !hasPermission(t.Tx, t.Uid, "DELETE", "user", t.UUID) {
+		t.Rw.WriteHeader(http.StatusForbidden)
+		return
+	}
+
 	result, err := t.Tx.Exec(`DELETE FROM "users" WHERE "id" = $1`, t.UUID)
 	if err != nil {
 		panic(err)
